requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
pydub
yt_dlp
